+++ features/fixtures/maze_runner/Assets/Plugins/iOS/native_code.mm
@@ -22,6 +22,8 @@ void ClearPersistentData() {
 }
 
 void RaiseCocoaSignal() {
+    NSLog(@"RaiseCocoaSignal");
+
     abort();
 }