+++ features/scripts/build_maze_runner.sh
@@ -37,31 +37,33 @@ pushd $SCRIPT_DIR
     echo "Expecting to find Bugsnag package in: $package_path"
   popd
   pushd ../fixtures
+    import_log_file="$package_path/unity_import.log"
     log_file="$package_path/unity.log"
     project_path="$(pwd)/maze_runner"
 
     # Run unity and immediately exit afterwards, log all output
-    DEFAULT_CLI_ARGS="-nographics -quit -batchmode -logFile $log_file"
+    DEFAULT_CLI_ARGS="-nographics -quit -batchmode"
 
-    if [[ "$UNITY_PATH" == *"2021"* ]]; then
-      # Copy the 2021 package manifest in for that version
-      mkdir -p maze_runner/Packages
-      cp manifests/manifest_2021.json maze_runner/Packages/manifest.json
-    elif [[ "$UNITY_PATH" == *"2020"* ]]; then
-      # Copy the 2020 package manifest in for that version
-      mkdir -p maze_runner/Packages
-      cp manifests/manifest_2020.json maze_runner/Packages/manifest.json
-    elif [[ "$UNITY_PATH" == *"2019"* ]]; then
-      # Copy the 2019 package manifest in for that version
-      mkdir -p maze_runner/Packages
-      cp manifests/manifest_2019.json maze_runner/Packages/manifest.json
-    else
-      # On all other versions disable the package manager
-      DEFAULT_CLI_ARGS="${DEFAULT_CLI_ARGS} -noUpm"
-    fi
+    # if [[ "$UNITY_PATH" == *"2021"* ]]; then
+    #   # Copy the 2021 package manifest in for that version
+    #   mkdir -p maze_runner/Packages
+    #   cp manifests/manifest_2021.json maze_runner/Packages/manifest.json
+    # elif [[ "$UNITY_PATH" == *"2020"* ]]; then
+    #   # Copy the 2020 package manifest in for that version
+    #   mkdir -p maze_runner/Packages
+    #   cp manifests/manifest_2020.json maze_runner/Packages/manifest.json
+    # elif [[ "$UNITY_PATH" == *"2019"* ]]; then
+    #   # Copy the 2019 package manifest in for that version
+    #   mkdir -p maze_runner/Packages
+    #   cp manifests/manifest_2019.json maze_runner/Packages/manifest.json
+    # else
+    #   # On all other versions disable the package manager
+    #   DEFAULT_CLI_ARGS="${DEFAULT_CLI_ARGS} -noUpm"
+    # fi
 
     echo "Importing $package_path/Bugsnag.unitypackage into $project_path"
     "$UNITY_PATH" $DEFAULT_CLI_ARGS \
+      -logFile $import_log_file \
       -projectPath $project_path \
       -ignoreCompilerErrors \
       -importPackage "$package_path/Bugsnag.unitypackage"
@@ -71,6 +73,7 @@ pushd $SCRIPT_DIR
     echo "Building the fixture for $PLATFORM"
 
     "$UNITY_PATH" $DEFAULT_CLI_ARGS \
+      -logFile $log_file \
       -projectPath $project_path \
       -executeMethod "Builder.$PLATFORM"
     RESULT=$?