+++ scripts/ci-build-macos-packages.sh
@@ -1,19 +1,8 @@
-#!/bin/bash
+#!/bin/bash -e
 
 # Build the MacOS and WebGL fixtures
 ./features/scripts/build_maze_runner.sh macos
-CODE=$?
-if [[ $CODE != 0 ]]; then
-  echo "Error, exit code: $CODE"
-  exit $CODE
-fi
-
 ./features/scripts/build_maze_runner.sh webgl
-CODE=$?
-if [[ $? != 0 ]]; then
-  echo "Error, exit code: $CODE"
-  exit $CODE
-fi
 
 pushd features/fixtures/maze_runner/build
   zip -r MacOS-$UNITY_VERSION.zip MacOS