+++ scripts/ci-run-webgl-tests.sh
@@ -1,7 +1,9 @@
-#!/bin/bash
+#!/bin/bash -e
+
 pushd features/fixtures/maze_runner/build
   unzip WebGL-$UNITY_VERSION.zip
 popd
 
 bundle install
-bundle exec maze-runner --farm=local --browser=firefox features/desktop
+# TODO: WebGL persistence tests are currently skipped pending PLAT-8151
+bundle exec maze-runner --farm=local --browser=firefox -e features/desktop/persistence.feature features/desktop