+++ src/Assets/Bugsnag/Scripts/BugsnagSettingsObject.cs
@@ -26,6 +26,8 @@ namespace BugsnagUnity
         public long LaunchDurationMillis = 5000;
         public int MaximumBreadcrumbs = 50;
         public int MaxPersistedEvents = 32;
+        public int MaxPersistedSessions = 128;
+        public int MaxReportedThreads = 200;
         public string NotifyEndpoint = "https://notify.bugsnag.com";
         public EditorLogLevel NotifyLogLevel = EditorLogLevel.Exception;
         public bool PersistUser = true;
@@ -79,6 +81,8 @@ namespace BugsnagUnity
             config.LaunchDurationMillis = LaunchDurationMillis;
             config.MaximumBreadcrumbs = MaximumBreadcrumbs;
             config.MaxPersistedEvents = MaxPersistedEvents;
+            config.MaxPersistedSessions = MaxPersistedSessions;
+            config.MaxReportedThreads = MaxReportedThreads;
             config.NotifyLogLevel = GetLogTypeFromLogLevel( NotifyLogLevel );
             config.SendThreads = SendThreads;
             if (!string.IsNullOrEmpty(NotifyEndpoint) && !string.IsNullOrEmpty(SessionEndpoint))