+++ src/BugsnagUnity/Client.cs
@@ -28,7 +28,7 @@ namespace BugsnagUnity
 
         private MaximumLogTypeCounter _logTypeCounter;
 
-        protected IDelivery Delivery => NativeClient.Delivery;
+        private Delivery _delivery;
 
         object CallbackLock { get; } = new object();
 
@@ -52,6 +52,8 @@ namespace BugsnagUnity
         public Client(INativeClient nativeClient)
         {
             NativeClient = nativeClient;
+            _delivery = new Delivery(nativeClient.Configuration);
+            FileManager.InitFileManager(nativeClient.Configuration);
             MainThread = Thread.CurrentThread;
             SessionTracking = new SessionTracker(this);
             InitStopwatches();
@@ -65,6 +67,7 @@ namespace BugsnagUnity
             InitInitialSessionCheck();          
             CheckForMisconfiguredEndpointsWarning();
             AddBugsnagLoadedBreadcrumb();
+            _delivery.StartDeliveringCachedPayloads();
         }
 
         private void InitFeatureFlags()
@@ -135,7 +138,7 @@ namespace BugsnagUnity
             else
             {
                 // otherwise create one
-                _cachedUser = new User { Id = SystemInfo.deviceUniqueIdentifier };
+                _cachedUser = new User { Id = FileManager.GetDeviceId() };
                 // see if a native user is avaliable
                 NativeClient.PopulateUser(_cachedUser);
             }
@@ -200,7 +203,7 @@ namespace BugsnagUnity
             {
                 return;
             }
-            Delivery.Send(payload);
+            _delivery.Deliver(payload);
         }
 
         void MultiThreadedNotify(string condition, string stackTrace, LogType logType)
@@ -426,31 +429,10 @@ namespace BugsnagUnity
                 // If the callback causes an exception, ignore it and execute the next one
             }
 
-
-            lock (CallbackLock)
-            {
-                foreach (var onSendErrorCallback in Configuration.GetOnSendErrorCallbacks())
-                {
-                    try
-                    {
-                        if (!onSendErrorCallback.Invoke(@event))
-                        {
-                            return;
-                        }
-                    }
-                    catch
-                    {
-                        // If the callback causes an exception, ignore it and execute the next one
-                    }
-                }
-            }
-
-            @event.RedactMetadata(Configuration);
-
             var report = new Report(Configuration, @event);
-
             if (!report.Ignored)
             {
+                FileManager.AddPendingPayload(report);
                 Send(report);
                 if (Configuration.IsBreadcrumbTypeEnabled(BreadcrumbType.Error))
                 {
@@ -506,6 +488,7 @@ namespace BugsnagUnity
                     }
                     _backgroundStopwatch.Reset();
                 }
+                _delivery.StartDeliveringCachedPayloads();
             }
             else
             {