+++ src/BugsnagUnity/MainThreadDispatchBehaviour.cs
@@ -81,30 +81,23 @@ namespace BugsnagUnity
             action.Invoke();
         }
 
-
-        private static MainThreadDispatchBehaviour _instance = null;
-
-        public static bool Exists()
-        {
-            return _instance != null;
-        }
+        private static MainThreadDispatchBehaviour _instance;
 
         public static MainThreadDispatchBehaviour Instance()
         {
-            if (!Exists())
+            if (_instance == null)
             {
-                throw new Exception("MainThreadDispatchBehaviour could not find the MainThreadDispatchBehaviour object.");
+                _instance = new GameObject("Bugsnag main thread dispatcher").AddComponent<MainThreadDispatchBehaviour>();
             }
             return _instance;
         }
-
 
         void Awake()
         {
             if (_instance == null)
             {
                 _instance = this;
-                DontDestroyOnLoad(this.gameObject);
+                DontDestroyOnLoad(gameObject);
             }
         }