+++ src/BugsnagUnity/Native/Android/NativeInterface.cs
@@ -275,9 +275,12 @@ namespace BugsnagUnity
             obj.Call("setContext", config.Context);
             obj.Call("setMaxBreadcrumbs", config.MaximumBreadcrumbs);
             obj.Call("setMaxPersistedEvents", config.MaxPersistedEvents);
+            obj.Call("setMaxPersistedSessions", config.MaxPersistedSessions);
             obj.Call("setPersistUser", config.PersistUser);
             obj.Call("setLaunchDurationMillis", config.LaunchDurationMillis);
             obj.Call("setSendLaunchCrashesSynchronously", config.SendLaunchCrashesSynchronously);
+            obj.Call("setMaxReportedThreads", config.MaxReportedThreads);
+
 
             if (config.GetUser() != null)
             {