+++ src/BugsnagUnity/Payload/IPayload.cs
@@ -7,8 +7,20 @@ namespace BugsnagUnity.Payload
 {
     public interface IPayload
     {
-        Uri Endpoint { get; }
+        Uri Endpoint { get; set; }
 
-        KeyValuePair<string, string>[] Headers { get; }
+        KeyValuePair<string, string>[] Headers { get; set; }
+
+        string Id { get; set; }
+
+        PayloadType PayloadType { get; }
+
+        Dictionary<string, object> GetSerialisablePayload();
+    }
+
+    public enum PayloadType
+    {
+        Session,
+        Event
     }
 }