+++ tests/UnityEngine/Object.cs
@@ -1,0 +1,15 @@
+﻿using System;
+namespace UnityEngine
+{
+    public class Object
+    {
+        public Object()
+        {
+        }
+
+        public static bool op_Equality(UnityEngine.Object a, UnityEngine.Object b)
+        {
+            return a == b;
+        }
+    }
+}