+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 6.2.0 (2022-03-16)
+
+### Enhancements
+
+* Added event, session and device id persistence for Windows and WebGL builds [#512](https://github.com/bugsnag/bugsnag-unity/pull/512) [#509](https://github.com/bugsnag/bugsnag-unity/pull/509) [#514](https://github.com/bugsnag/bugsnag-unity/pull/514)
+
+* Add `Configuration.MaxReportedThreads` config option to set the [native Android option](https://docs.bugsnag.com/platforms/android/configuration-options/#maxreportedthreads) [523](https://github.com/bugsnag/bugsnag-unity/pull/523)
+
+* Update bugsnag-android to v5.20.0
+  * The number of threads reported can now be limited using `Configuration.setMaxReportedThreads` (defaulting to 200)
+    [bugsnag-android#1607](https://github.com/bugsnag/bugsnag-android/pull/1607)
+  * Improved the performance and stability of the NDK and ANR plugins by caching JNI references on start
+    [bugsnag-android#1596](https://github.com/bugsnag/bugsnag-android/pull/1596)
+    [bugsnag-android#1601](https://github.com/bugsnag/bugsnag-android/pull/1601)
+
 ## 6.1.0 (2022-02-08)
 
 ### Enhancements